Researchers Exchange Participation
A Java web application called Researchers Exchange Participations, using
the MVC pattern. REP is a platform for researchers to exchange
participations. Moreover, recruiting serious participants has been
always a burden on a researcher's shoulder, thus, this platform gives
researchers the opportunity to do that effectively and in a suitable
time.